Restructure engine samples into directories (#244)
* Restructure engine samples into directories
Remove asset_renderer sample, no longer relevant

* Fix cubinhos generate cmake module

* Modernize scene sample

* Move initialize logger into Cubos

* Remove unnecessary namespace
