+++ engine/samples/custom-asset/assets/sample.meta
