+++ engine/samples/custom-asset/assets/sample.txt
