+++ engine/samples/custom-asset/main.cpp
