+++ engine/samples/scene/components/num.hpp
@@ -1,0 +1,11 @@
+#ifndef COMPONENTS_NUM_HPP
+#define COMPONENTS_NUM_HPP
+
+#include <components/base.hpp>
+
+struct [[cubos::component("num", VecStorage)]] Num
+{
+    int value;
+};
+
+#endif // COMPONENTS_NUM_HPP
