+++ engine/samples/ssao/assets/cubos.meta
