+++ engine/samples/ssao/assets/cubos.pal
